Renamed project from 'contacts' to 'contacts_plugin'
