+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -19,6 +19,6 @@
 ### :thinking: Checklist before submitting
 
 - [ ] All projects build
-- [ ] Follows style guide lines ([code style guide](https://github.com/BaseflowIT/flutter-contacts/blob/develop/CONTRIBUTING.md))
+- [ ] Follows style guide lines ([code style guide](https://github.com/BaseflowIT/flutter-contacts-plugin/blob/develop/CONTRIBUTING.md))
 - [ ] Relevant documentation was updated
 - [ ] Rebased onto current develop
