+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'contacts'
+rootProject.name = 'contacts_plugin'
