+++ example/android/app/src/main/kotlin/com/example/contactspluginexample/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.contactsexample
+package com.example.contactspluginexample
 
 import android.os.Bundle
