+++ example/ios/Podfile
@@ -50,9 +50,9 @@ target 'Runner' do
   # Plugin Pods
   plugin_pods = parse_KV_file('../.flutter-plugins')
   plugin_pods.map { |p|
-    symlink = File.join('.symlinks', 'plugins', 'contacts')
+    symlink = File.join('.symlinks', 'plugins', p[:name])
     File.symlink(p[:path], symlink)
-    pod 'BFContacts', :path => File.join(symlink, 'ios')
+    pod p[:name], :path => File.join(symlink, 'ios')
   }
 end
 
@@ -60,7 +60,6 @@ post_install do |installer|
   installer.pods_project.targets.each do |target|
     target.build_configurations.each do |config|
       config.build_settings['ENABLE_BITCODE'] = 'NO'
-      config.build_settings['SWIFT_VERSION'] = '4.0'
     end
   end
 end
