+++ example/pubspec.yaml
@@ -1,5 +1,5 @@
-name: contacts_example
-description: Demonstrates how to use the contacts plugin.
+name: contacts_plugin_example
+description: Demonstrates how to use the contacts_plugin plugin.
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -24,7 +24,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  contacts:
+  contacts_plugin:
     path: ../
 
 # For information on the generic Dart part of this file, see the
