+++ ios/Classes/SwiftContactsPlugin.swift
@@ -3,7 +3,7 @@ import UIKit
     
 public class SwiftContactsPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "contacts", binaryMessenger: registrar.messenger())
+    let channel = FlutterMethodChannel(name: "contacts_plugin", binaryMessenger: registrar.messenger())
     let instance = SwiftContactsPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }
