+++ ios/Classes/strategies/ContactStoreStrategy.swift
@@ -5,12 +5,13 @@
 //  Created by Maurits van Beusekom on 28/08/2018.
 //
 
+import Contacts
 import Foundation
 
 @available(iOS 9.0, *)
 class ContactStoreStrategy : ContactStrategy {
     func fetchContacts() -> [Contact] {
-        return [Contact()]
+        return [Contact]()
     }
     
 }
