+++ lib/models/email_address.dart
@@ -1,4 +1,4 @@
-part of contacts;
+part of contacts_plugin;
 
 class EmailAddress {
   String address;
