+++ lib/models/instant_messaging_account.dart
@@ -1,4 +1,4 @@
-part of contacts;
+part of contacts_plugin;
 
 class InstantMessagingAccount {
   String account;
