+++ lib/models/organization.dart
@@ -1,4 +1,4 @@
-part of contacts;
+part of contacts_plugin;
 
 class Organization {
   String contactTitle;
