+++ lib/models/phone_number.dart
@@ -1,4 +1,4 @@
-part of contacts;
+part of contacts_plugin;
 
 class PhoneNumber {
   String label;
