+++ lib/models/relationship.dart
@@ -1,4 +1,4 @@
-part of contacts;
+part of contacts_plugin;
 
 class Relationship {
   String name;
