+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: contacts
+name: contacts_plugin
 description: A new flutter plugin project.
 version: 0.0.1
 author:
@@ -11,17 +11,13 @@ dependencies:
   flutter:
     sdk: flutter
 
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-    
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
 
 # The following section is specific to Flutter.
 flutter:
   plugin:
-    androidPackage: com.example.contacts
+    androidPackage: com.example.contactsplugin
     pluginClass: ContactsPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
